Improve react sample
