+++ typescript/native-auth/react-nextjs-sample/src/app/reset-password/styles/styles.ts
@@ -35,4 +35,13 @@ export const styles = {
         marginTop: "10px",
         fontSize: "14px",
     },
+    h2: {
+        margin: "0 0 10px 0",
+    },
+    signed_in_msg: {
+        padding: "20px",
+        border: "1px solid #ddd",
+        borderRadius: "4px",
+        marginTop: "20px",
+    }
 };
