+++ typescript/native-auth/react-nextjs-sample/src/app/sign-in/styles/styles.ts
@@ -30,4 +30,7 @@ export const styles = {
         color: "#d13438",
         marginTop: "10px",
     },
+    h2: {
+        margin: "0 0 10px 0",
+    },
 } as const;
