--- typescript/native-auth/react-nextjs-sample/src/app/sign-up-attribute/styles/styles.ts
@@ -1,33 +1,0 @@
-export const styles = {
-    container: {
-        maxWidth: "500px", // Increased width for additional fields
-        margin: "40px auto",
-        padding: "20px",
-        border: "1px solid #ccc",
-        borderRadius: "8px",
-    },
-    form: {
-        display: "flex",
-        flexDirection: "column",
-        gap: "15px",
-    },
-    input: {
-        padding: "8px",
-        border: "1px solid #ccc",
-        borderRadius: "4px",
-        fontSize: "16px",
-    },
-    button: {
-        padding: "10px",
-        backgroundColor: "#0078d4",
-        color: "white",
-        border: "none",
-        borderRadius: "4px",
-        cursor: "pointer",
-        fontSize: "16px",
-    },
-    error: {
-        color: "#d13438",
-        marginTop: "10px",
-    },
-} as const;
