--- typescript/native-auth/react-nextjs-sample/src/app/sign-up-password/components/InitialFormWithPassword.tsx
@@ -1,67 +1,0 @@
-import { styles } from "../styles/styles";
-
-interface InitialFormWithPasswordProps {
-    onSubmit: (e: React.FormEvent) => Promise<void>;
-    firstName: string;
-    setFirstName: (value: string) => void;
-    lastName: string;
-    setLastName: (value: string) => void;
-    email: string;
-    setEmail: (value: string) => void;
-    password: string;
-    setPassword: (value: string) => void;
-    loading: boolean;
-}
-
-export function InitialFormWithPassword({
-    onSubmit,
-    firstName,
-    setFirstName,
-    lastName,
-    setLastName,
-    email,
-    setEmail,
-    password,
-    setPassword,
-    loading,
-}: InitialFormWithPasswordProps) {
-    return (
-        <form onSubmit={onSubmit} style={styles.form}>
-            <input
-                type="text"
-                placeholder="First Name"
-                value={firstName}
-                onChange={(e) => setFirstName(e.target.value)}
-                style={styles.input}
-                required
-            />
-            <input
-                type="text"
-                placeholder="Last Name"
-                value={lastName}
-                onChange={(e) => setLastName(e.target.value)}
-                style={styles.input}
-                required
-            />
-            <input
-                type="email"
-                placeholder="Email"
-                value={email}
-                onChange={(e) => setEmail(e.target.value)}
-                style={styles.input}
-                required
-            />
-            <input
-                type="password"
-                placeholder="Password"
-                value={password}
-                onChange={(e) => setPassword(e.target.value)}
-                style={styles.input}
-                required
-            />
-            <button type="submit" style={styles.button} disabled={loading}>
-                {loading ? "Signing up..." : "Sign Up"}
-            </button>
-        </form>
-    );
-}
