--- typescript/native-auth/react-nextjs-sample/src/app/sign-up-password/components/SignUpResult.tsx
@@ -1,34 +1,0 @@
-import { useRouter } from "next/navigation";
-
-export function SignUpResultPage() {
-    const router = useRouter();
-
-    return (
-        <div
-            style={{
-                padding: "20px",
-                border: "1px solid #ddd",
-                borderRadius: "4px",
-                marginTop: "20px",
-            }}
-        >
-            <div>Sign Up Successful</div>
-            <div style={{ marginTop: "20px" }}>
-                <button
-                    style={{
-                        padding: "10px",
-                        backgroundColor: "#0078d4",
-                        color: "white",
-                        border: "none",
-                        borderRadius: "4px",
-                        cursor: "pointer",
-                        fontSize: "16px",
-                    }}
-                    onClick={() => router.push("/sign-in")}
-                >
-                    Go to Sign In
-                </button>
-            </div>
-        </div>
-    );
-}
