--- typescript/native-auth/react-nextjs-sample/src/app/sign-up/components/InitialForm.tsx
@@ -1,46 +1,0 @@
-import { styles } from "../styles/styles";
-
-interface InitialFormProps {
-    onSubmit: (e: React.FormEvent) => Promise<void>;
-    firstName: string;
-    setFirstName: (value: string) => void;
-    lastName: string;
-    setLastName: (value: string) => void;
-    email: string;
-    setEmail: (value: string) => void;
-    loading: boolean;
-}
-
-export function InitialForm({ onSubmit, firstName, setFirstName, lastName, setLastName, email, setEmail, loading }: InitialFormProps) {
-    return (
-        <form onSubmit={onSubmit} style={styles.form}>
-            <input
-                type="text"
-                placeholder="First Name"
-                value={firstName}
-                onChange={(e) => setFirstName(e.target.value)}
-                style={styles.input}
-                required
-            />
-            <input
-                type="text"
-                placeholder="Last Name"
-                value={lastName}
-                onChange={(e) => setLastName(e.target.value)}
-                style={styles.input}
-                required
-            />
-            <input
-                type="email"
-                placeholder="Email"
-                value={email}
-                onChange={(e) => setEmail(e.target.value)}
-                style={styles.input}
-                required
-            />
-            <button type="submit" style={styles.button} disabled={loading}>
-                {loading ? "Signing up..." : "Sign Up"}
-            </button>
-        </form>
-    );
-}
