+++ typescript/native-auth/react-nextjs-sample/src/app/sign-up/styles/styles.ts
@@ -30,4 +30,13 @@ export const styles = {
         color: "#d13438",
         marginTop: "10px",
     },
+    h2: {
+        margin: "0 0 10px 0",
+    },
+    signed_in_msg: {
+        padding: "20px",
+        border: "1px solid #ddd",
+        borderRadius: "4px",
+        marginTop: "20px",
+    }
 } as const;
